+++ ElvUI_LuckyoneUI/Modules/Auras/Filters.lua
@@ -15,15 +15,15 @@ local function Add(list, ids, value)
 	end
 end
 
--- General vars
-local unitframe = E.global['unitframe'] or {}
-local aurafilters = unitframe['aurafilters'] or {}
-local aurawatch = unitframe['aurawatch'] or {}
-local auraHighlight = unitframe['AuraHighlightColors'] or {}
-
 -- Aura filters: Retail
 function Private:Setup_Filters_Retail()
 	if not E.Retail then return end
+
+	-- General vars
+	local unitframe = E.global['unitframe'] or {}
+	local aurafilters = unitframe['aurafilters'] or {}
+	local aurawatch = unitframe['aurawatch'] or {}
+	local auraHighlight = unitframe['AuraHighlightColors'] or {}
 
 	-- Classes setup
 	local classes = {
@@ -46,7 +46,7 @@ function Private:Setup_Filters_Retail()
 		whitelist = {},
 		-- Healers
 		DRUID = { 207386, 188550, 155777, 203554, 774, 102351, 102352, 8936, 33763, 48438, 200389 },
-		EVOKER = { 367364, 376788, 363502, 406732, 410089, 364343, 395152, 373267, 366155, 369459, 357170, 360827, 355941, 406789 },
+		EVOKER = { 367364, 376788, 363502, 406732, 410089, 364343, 395152, 373267, 366155, 369459, 357170, 360827, 355941, 406789, 361022 },
 		MONK = { 325209, 115175, 191840, 116849, 119611, 124682, 116841 },
 		PALADIN = { 1044, 200025, 287280, 148039, 6940, 156910, 157047, 204018, 223306, 53563, 1022, 395180 },
 		PRIEST = { 139, 193065, 10060, 214206, 17, 33206, 194384, 41635, 47788, 6788, 77489 },
@@ -151,6 +151,11 @@ function Private:Setup_Filters_Retail()
 		["point"] = "TOP",
 		["displayText"] = true,
 		["sizeOffset"] = 4,
+		["style"] = "texturedIcon",
+	}
+	classes['EVOKER'][361022] = {
+		["point"] = "BOTTOMLEFT",
+		["sizeOffset"] = 6,
 		["style"] = "texturedIcon",
 	}
 
@@ -259,6 +264,12 @@ end
 -- Aura filters: Cataclysm
 function Private:Setup_Filters_Cata()
 	if not E.Cata then return end
+
+	-- General vars
+	local unitframe = E.global['unitframe'] or {}
+	local aurafilters = unitframe['aurafilters'] or {}
+	local aurawatch = unitframe['aurawatch'] or {}
+	local auraHighlight = unitframe['AuraHighlightColors'] or {}
 
 	-- Classes setup
 	local classes = {
@@ -545,6 +556,12 @@ end
 function Private:Setup_Filters_Classic()
 	if not E.Classic then return end
 
+	-- General vars
+	local unitframe = E.global['unitframe'] or {}
+	local aurafilters = unitframe['aurafilters'] or {}
+	local aurawatch = unitframe['aurawatch'] or {}
+	local auraHighlight = unitframe['AuraHighlightColors'] or {}
+
 	-- Classes setup
 	local classes = {
 		-- Healers
@@ -565,16 +582,224 @@ function Private:Setup_Filters_Classic()
 		blacklist = {},
 		whitelist = {},
 		-- Healers
-		DRUID = {},
-		PALADIN = {},
-		PRIEST = {},
-		SHAMAN = {},
+		DRUID = { 21849, 467, 1126, 8936, 408120, 774, 29166 },
+		PALADIN = { 19740, 25894, 1044, 25782, 6940, 19746, 1022, 19742, 19977, 465 },
+		PRIEST = { 6346, 139, 27683, 1243, 10060, 402004, 17, 27681, 14752, 401877, 21562, 976 },
+		SHAMAN = { 8072, 25909, 10596, 8182, 29203, 8185, 16237, 16191, 5677, 5672 },
 		-- Others
-		MAGE = {},
-		WARLOCK = {},
-		HUNTER = {},
+		MAGE = { 1008, 604, 1459, 23028, 130, 400735 },
+		WARLOCK = { 2970, 6512, 11743, 5697 },
+		HUNTER = { 19506, 13159, 20043 },
 		ROGUE = {},
-		WARRIOR = {}
+		WARRIOR = { 6673 }
+	}
+
+	-- Druid
+	classes["DRUID"][21849] = {
+		["style"] = "texturedIcon",
+	}
+	classes["DRUID"][467] = {
+		["style"] = "texturedIcon",
+	}
+	classes["DRUID"][1126] = {
+		["style"] = "texturedIcon",
+	}
+	classes["DRUID"][8936] = {
+		["style"] = "texturedIcon",
+	}
+	classes["DRUID"][408120] = {
+		["style"] = "texturedIcon",
+	}
+	classes["DRUID"][774] = {
+		["style"] = "texturedIcon",
+	}
+	classes["DRUID"][29166] = {
+		["style"] = "texturedIcon",
+	}
+
+	-- Paladin
+	classes["PALADIN"][19740] = {
+		["style"] = "texturedIcon",
+	}
+	classes["PALADIN"][25894] = {
+		["style"] = "texturedIcon",
+	}
+	classes["PALADIN"][1044] = {
+		["style"] = "texturedIcon",
+	}
+	classes["PALADIN"][25782] = {
+		["style"] = "texturedIcon",
+	}
+	classes["PALADIN"][6940] = {
+		["style"] = "texturedIcon",
+	}
+	classes["PALADIN"][19746] = {
+		["style"] = "texturedIcon",
+	}
+	classes["PALADIN"][1022] = {
+		["style"] = "texturedIcon",
+	}
+	classes["PALADIN"][19742] = {
+		["style"] = "texturedIcon",
+	}
+	classes["PALADIN"][19977] = {
+		["style"] = "texturedIcon",
+	}
+	classes["PALADIN"][465] = {
+		["style"] = "texturedIcon",
+	}
+
+	-- Priest
+	classes["PRIEST"][6346] = {
+		["point"] = "BOTTOMRIGHT",
+		["yOffset"] = -1,
+		["anyUnit"] = true,
+		["style"] = "texturedIcon",
+		["xOffset"] = 1,
+	}
+	classes["PRIEST"][139] = {
+		["point"] = "TOPLEFT",
+		["yOffset"] = 1,
+		["style"] = "texturedIcon",
+		["xOffset"] = -1,
+	}
+	classes["PRIEST"][27683] = {
+		["yOffset"] = -1,
+		["style"] = "texturedIcon",
+		["xOffset"] = 25,
+	}
+	classes["PRIEST"][1243] = {
+		["point"] = "BOTTOMLEFT",
+		["yOffset"] = -1,
+		["style"] = "texturedIcon",
+		["xOffset"] = -1,
+	}
+	classes["PRIEST"][10060] = {
+		["yOffset"] = 1,
+		["style"] = "texturedIcon",
+	}
+	classes["PRIEST"][402004] = {
+		["yOffset"] = 1,
+		["style"] = "texturedIcon",
+	}
+	classes["PRIEST"][17] = {
+		["point"] = "TOPLEFT",
+		["yOffset"] = 1,
+		["style"] = "texturedIcon",
+		["xOffset"] = 25,
+	}
+	classes["PRIEST"][27681] = {
+		["point"] = "BOTTOMLEFT",
+		["yOffset"] = -1,
+		["style"] = "texturedIcon",
+		["xOffset"] = 12,
+	}
+	classes["PRIEST"][14752] = {
+		["point"] = "BOTTOMLEFT",
+		["yOffset"] = -1,
+		["style"] = "texturedIcon",
+		["xOffset"] = 12,
+	}
+	classes["PRIEST"][401877] = {
+		["point"] = "TOPLEFT",
+		["yOffset"] = 1,
+		["countY"] = 0,
+		["style"] = "texturedIcon",
+		["countX"] = 0,
+		["xOffset"] = 12,
+	}
+	classes["PRIEST"][21562] = {
+		["point"] = "BOTTOMLEFT",
+		["yOffset"] = -1,
+		["style"] = "texturedIcon",
+		["xOffset"] = -1,
+	}
+	classes["PRIEST"][976] = {
+		["yOffset"] = -1,
+		["style"] = "texturedIcon",
+		["xOffset"] = 25,
+	}
+
+	-- Shaman
+	classes["SHAMAN"][8072] = {
+		["style"] = "texturedIcon",
+	}
+	classes["SHAMAN"][25909] = {
+		["style"] = "texturedIcon",
+	}
+	classes["SHAMAN"][10596] = {
+		["style"] = "texturedIcon",
+	}
+	classes["SHAMAN"][8182] = {
+		["style"] = "texturedIcon",
+	}
+	classes["SHAMAN"][29203] = {
+		["style"] = "texturedIcon",
+	}
+	classes["SHAMAN"][8185] = {
+		["style"] = "texturedIcon",
+	}
+	classes["SHAMAN"][16237] = {
+		["style"] = "texturedIcon",
+	}
+	classes["SHAMAN"][16191] = {
+		["style"] = "texturedIcon",
+	}
+	classes["SHAMAN"][5677] = {
+		["style"] = "texturedIcon",
+	}
+	classes["SHAMAN"][5672] = {
+		["style"] = "texturedIcon",
+	}
+
+	-- Mage
+	classes["MAGE"][1008] = {
+		["style"] = "texturedIcon",
+	}
+	classes["MAGE"][604] = {
+		["style"] = "texturedIcon",
+	}
+	classes["MAGE"][1459] = {
+		["style"] = "texturedIcon",
+	}
+	classes["MAGE"][23028] = {
+		["style"] = "texturedIcon",
+	}
+	classes["MAGE"][130] = {
+		["style"] = "texturedIcon",
+	}
+	classes["MAGE"][400735] = {
+		["style"] = "texturedIcon",
+	}
+
+	-- Warlock
+	classes["WARLOCK"][2970] = {
+		["style"] = "texturedIcon",
+	}
+	classes["WARLOCK"][6512] = {
+		["style"] = "texturedIcon",
+	}
+	classes["WARLOCK"][11743] = {
+		["style"] = "texturedIcon",
+	}
+	classes["WARLOCK"][5697] = {
+		["style"] = "texturedIcon",
+	}
+
+	-- Hunter
+	classes["HUNTER"][19506] = {
+		["style"] = "texturedIcon",
+	}
+	classes["HUNTER"][13159] = {
+		["style"] = "texturedIcon",
+	}
+	classes["HUNTER"][20043] = {
+		["style"] = "texturedIcon",
+	}
+
+	-- Warrior
+	classes["WARRIOR"][6673] = {
+		["style"] = "texturedIcon",
 	}
 
 	for class, classIDs in pairs(ids) do